jwt2/requirements.txt
orjson
pyjwt[crypto]